refactored code according to comments
